+++ lib/src/services/scheduler/schedule.dart
@@ -3,17 +3,30 @@
 abstract class Schedule {
   const Schedule();
 
-  factory Schedule.repeat(Duration interval) => RepeatSchedule(interval);
+  factory Schedule.repeat(Duration interval, {bool startNow = true}) =>
+      RepeatSchedule(interval, startNow);
 
   factory Schedule.daily(int hour, [int minute = 0, int second = 0]) =>
       DailySchedule(hour, minute, second);
+
+  DateTime findNext(DateTime? lastExecution);
 }
 
 /// Repeated execution with interval [Duration].
 class RepeatSchedule extends Schedule {
   final Duration interval;
+  final bool startNow;
 
-  const RepeatSchedule(this.interval);
+  const RepeatSchedule(this.interval, this.startNow);
+
+  @override
+  DateTime findNext(DateTime? execution) {
+    if (execution == null && startNow) {
+      return DateTime.now();
+    }
+
+    return (execution ?? DateTime.now()).add(interval);
+  }
 }
 
 class DailySchedule extends Schedule {
@@ -22,4 +35,20 @@ class DailySchedule extends Schedule {
   final int second;
 
   DailySchedule(this.hour, [this.minute = 0, this.second = 0]);
+
+  @override
+  DateTime findNext(DateTime? execution) {
+    if (execution != null) {
+      return execution.add(const Duration(days: 1));
+    } else {
+      final now = DateTime.now();
+      final today =
+          DateTime(now.year, now.month, now.day, hour, minute, second);
+      if (today.isAfter(now)) {
+        return today;
+      } else {
+        return today.add(const Duration(days: 1));
+      }
+    }
+  }
 }